Change distance to integer: user-friendly message.

+++ Spotr/src/com/csun/spotr/LocalMapViewActivity.java
@@ -118,7 +118,7 @@ public class LocalMapViewActivity extends MapActivity {
 					mapController.animateTo(geoFriend);
 					mapController.setZoom(19);
 					mapView.invalidate();
-					Toast.makeText(LocalMapViewActivity.this, d + " m", Toast.LENGTH_SHORT).show();
+					Toast.makeText(LocalMapViewActivity.this, (int) d + " m", Toast.LENGTH_SHORT).show();
 					indexFriend++;
 				}
 			}
@@ -150,7 +150,7 @@ public class LocalMapViewActivity extends MapActivity {
 					mapController.animateTo(geoPlace);
 					mapController.setZoom(19);
 					mapView.invalidate();
-					Toast.makeText(LocalMapViewActivity.this, d + " m", Toast.LENGTH_SHORT).show();
+					Toast.makeText(LocalMapViewActivity.this, (int) d + " m", Toast.LENGTH_SHORT).show();
 					indexPlace++;
 				}
 			}
